+++ crates/radiate-python/src/bindings/genome/phenotype.rs
@@ -98,7 +98,7 @@ macro_rules! impl_from_py_phenotype {
 }
 
 impl_from_py_phenotype!(FloatChromosome);
-impl_from_py_phenotype!(IntChromosome<i32>);
+impl_from_py_phenotype!(IntChromosome<i64>);
 impl_from_py_phenotype!(BitChromosome);
 impl_from_py_phenotype!(CharChromosome);
 impl_from_py_phenotype!(GraphChromosome<Op<f32>>);
